Split calls.rs and calls_p4.rs at clean impl-block boundaries to clear max-lines
