+++ crates/almide-mir/src/lower/calls_p4.rs
@@ -608,1246 +608,4 @@ impl LowerCtx {
         }
         self.lower_scalar_binop_eq_list_map(op, left, right)
     }
-
-    /// Extracted from `Self::lower_scalar_binop_eq_family` (eighth-round split, cog
-    /// reduction): the String/Value deep-equality sub-chain, verbatim.
-    fn lower_scalar_binop_eq_string_value(
-        &mut self,
-        op: &almide_ir::BinOp,
-        left: &IrExpr,
-        right: &IrExpr,
-    ) -> Option<ValueId> {
-        use almide_ir::BinOp;
-        // STRING equality (`c == ":"` / `a != b` over String) → the self-host
-        // `string.eq` byte-compare call (→ scalar Bool). Both operands are BORROWED
-        // heap String handles (the call reads + copies; no ownership event). `!=` is
-        // `1 - eq`. This is the dominant real-parser condition; without it the cond
-        // silently lowered to 0 (false) — the yaml/char-scan miscompile.
-        if matches!(op, BinOp::Eq | BinOp::Neq) && matches!(left.ty, Ty::String) {
-            let args = [left.clone(), right.clone()];
-            let eq = self
-                .lower_pure_module_value_call("string", "eq", &args, &Ty::Bool)
-                .ok()?;
-            if matches!(op, BinOp::Eq) {
-                return Some(eq);
-            }
-            let one = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst: one, value: 1 });
-            let dst = self.fresh_value();
-            self.ops.push(Op::IntBinOp { dst, op: crate::IntOp::Sub, a: one, b: eq });
-            return Some(dst);
-        }
-        // `value.eq` deep-structural call (→ scalar Bool) for a `Value == Value` / `!=`. Without
-        // this the heap `==` did not lower to a scalar cond, so an `if value==value …` fell to the
-        // both-arms linearization and ran BOTH arms (silent miscompile). Both operands BORROWED
-        // (value_eq only reads). `!=` is `1 - eq`. The recursive value_eq byte-matches v0's Value
-        // PartialEq.
-        if matches!(op, BinOp::Eq | BinOp::Neq) && crate::lower::is_value_ty(&left.ty) {
-            let args = [left.clone(), right.clone()];
-            let eq = self
-                .lower_pure_module_value_call("value", "eq", &args, &Ty::Bool)
-                .ok()?;
-            if matches!(op, BinOp::Eq) {
-                return Some(eq);
-            }
-            let one = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst: one, value: 1 });
-            let dst = self.fresh_value();
-            self.ops.push(Op::IntBinOp { dst, op: crate::IntOp::Sub, a: one, b: eq });
-            return Some(dst);
-        }
-        None
-    }
-
-    /// Extracted from `Self::lower_scalar_binop_eq_family` (eighth-round split, cog
-    /// reduction): the List/Map/Set deep-equality sub-chain, verbatim.
-    fn lower_scalar_binop_eq_list_map(
-        &mut self,
-        op: &almide_ir::BinOp,
-        left: &IrExpr,
-        right: &IrExpr,
-    ) -> Option<ValueId> {
-        if let Some(dst) = self.lower_scalar_binop_eq_list(op, left, right) {
-            return Some(dst);
-        }
-        self.lower_scalar_binop_eq_map_set(op, left, right)
-    }
-
-    /// Extracted from `Self::lower_scalar_binop_eq_list_map` (ninth-round split, cog
-    /// reduction): the List deep-equality sub-chain, verbatim.
-    /// Extracted from `Self::lower_scalar_binop_eq_list` (tenth-round split, cog
-    /// reduction): the element-type → `list.eq_*` callee-name lookup, verbatim (a static
-    /// value computation, no `&mut self` needed).
-    fn list_eq_call_variant(es: &[Ty]) -> Option<&'static str> {
-        if es.len() != 1 {
-            None
-        } else if matches!(es[0], Ty::Int) {
-            Some("eq_int")
-        } else if matches!(es[0], Ty::String) {
-            Some("eq_str")
-        } else if crate::lower::is_value_ty(&es[0]) {
-            Some("eq_value")
-        } else if matches!(es[0], Ty::Float) {
-            Some("eq_float")
-        } else if matches!(es[0], Ty::Bool) {
-            Some("eq_bool")
-        } else {
-            None
-        }
-    }
-
-    fn lower_scalar_binop_eq_list(
-        &mut self,
-        op: &almide_ir::BinOp,
-        left: &IrExpr,
-        right: &IrExpr,
-    ) -> Option<ValueId> {
-        use almide_ir::BinOp;
-        // `list_a == list_b` over a List[Int|String|Value]: a deep element-wise compare call
-        // (→ scalar Bool). Same both-arms-linearization fix as Value/String ==. element type
-        // picks the variant; other element types stay unhandled (the if then walls, loud).
-        if matches!(op, BinOp::Eq | BinOp::Neq) {
-            if let Ty::Applied(almide_lang::types::constructor::TypeConstructorId::List, es) =
-                &left.ty
-            {
-                let variant = Self::list_eq_call_variant(es);
-                if let Some(v) = variant {
-                    let args = [left.clone(), right.clone()];
-                    let eq = self
-                        .lower_pure_module_value_call("list", v, &args, &Ty::Bool)
-                        .ok()?;
-                    if matches!(op, BinOp::Eq) {
-                        return Some(eq);
-                    }
-                    let one = self.fresh_value();
-                    self.ops.push(Op::ConstInt { dst: one, value: 1 });
-                    let dst = self.fresh_value();
-                    self.ops.push(Op::IntBinOp { dst, op: crate::IntOp::Sub, a: one, b: eq });
-                    return Some(dst);
-                }
-            }
-        }
-        None
-    }
-
-    /// Extracted from `Self::lower_scalar_binop_eq_list_map` (ninth-round split, cog
-    /// reduction): the Map/Set deep-equality sub-chain, verbatim.
-    fn lower_scalar_binop_eq_map_set(
-        &mut self,
-        op: &almide_ir::BinOp,
-        left: &IrExpr,
-        right: &IrExpr,
-    ) -> Option<ValueId> {
-        use almide_ir::BinOp;
-        // `map_a == map_b` over the two implemented map reprs — a deep
-        // order-independent compare call (→ scalar Bool), same shape as list ==.
-        if matches!(op, BinOp::Eq | BinOp::Neq) {
-            let is_set_str = matches!(&left.ty,
-                Ty::Applied(almide_lang::types::constructor::TypeConstructorId::Set, a)
-                    if a.len() == 1 && matches!(a[0], Ty::String));
-            let is_map_skv = matches!(&left.ty,
-                Ty::Applied(almide_lang::types::constructor::TypeConstructorId::Map, a)
-                    if a.len() == 2 && matches!(a[0], Ty::String) && !crate::lower::is_heap_ty(&a[1]));
-            let admitted = crate::lower::is_map_ivh_ty(&left.ty)
-                || crate::lower::is_map_hval_ty(&left.ty)
-                || is_map_skv
-                || is_set_str;
-            if admitted {
-                let module = if is_set_str { "set" } else { "map" };
-                // Pass the BARE "eq" — `list_heap_call_name` attaches the repr
-                // suffix (`map.eq_ivh` / `map.eq_hval`) from the subject type,
-                // exactly like every other map call site.
-                let args = [left.clone(), right.clone()];
-                let eq = self
-                    .lower_pure_module_value_call(module, "eq", &args, &Ty::Bool)
-                    .ok()?;
-                if matches!(op, BinOp::Eq) {
-                    return Some(eq);
-                }
-                let one = self.fresh_value();
-                self.ops.push(Op::ConstInt { dst: one, value: 1 });
-                let dst = self.fresh_value();
-                self.ops.push(Op::IntBinOp { dst, op: crate::IntOp::Sub, a: one, b: eq });
-                return Some(dst);
-            }
-        }
-        None
-    }
-
-    /// Extracted from `Self::lower_scalar_binop` (seventh-round split, cog reduction):
-    /// the String-ordering-cmp + heap-typed `==`/`!=` sub-chain, verbatim.
-    fn lower_scalar_binop_cmp_and_heap_eq(
-        &mut self,
-        op: &almide_ir::BinOp,
-        left: &IrExpr,
-        right: &IrExpr,
-    ) -> Option<ValueId> {
-        use almide_ir::BinOp;
-        // String ordering `< <= > >=` → `string.cmp(a,b)` (lexicographic, -1/0/1) compared with
-        // 0. WITHOUT this the comparison fell through to the i64-handle path → arbitrary order
-        // (silent), or the if linearized both arms. Both operands BORROWED (cmp only reads).
-        if matches!(op, BinOp::Lt | BinOp::Lte | BinOp::Gt | BinOp::Gte)
-            && matches!(left.ty, Ty::String)
-        {
-            let args = [left.clone(), right.clone()];
-            let cmp = self
-                .lower_pure_module_value_call("string", "cmp", &args, &Ty::Int)
-                .ok()?;
-            let zero = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst: zero, value: 0 });
-            let iop = match op {
-                BinOp::Lt => crate::IntOp::Lt,
-                BinOp::Lte => crate::IntOp::Le,
-                BinOp::Gt => crate::IntOp::Gt,
-                _ => crate::IntOp::Ge,
-            };
-            let dst = self.fresh_value();
-            self.ops.push(Op::IntBinOp { dst, op: iop, a: cmp, b: zero });
-            return Some(dst);
-        }
-        // Heap `==` / `!=` in a VALUE position (Option/Result/tuple/record/custom variant —
-        // any layout the recursive typed-eq engine composes): the SAME materialized engine
-        // the unit-if cond uses. Operands materialize (a tracked Var borrowed, a fresh
-        // ctor/call an owned temp freed at frame teardown); the eq only reads. Was both-arms-
-        // linearized (silent). Rolls back fully on a shape outside the engine — the caller
-        // then defers/walls (loud, never wrong).
-        if matches!(op, BinOp::Eq | BinOp::Neq) && is_heap_ty(&left.ty) {
-            let ops_mark = self.ops.len();
-            let lhh_mark = self.live_heap_handles.len();
-            if let Some(eq) = self.lower_heap_eq_typed_materialized(left, right, &left.ty) {
-                if matches!(op, BinOp::Eq) {
-                    return Some(eq);
-                }
-                let one = self.fresh_value();
-                self.ops.push(Op::ConstInt { dst: one, value: 1 });
-                let dst = self.fresh_value();
-                self.ops.push(Op::IntBinOp { dst, op: crate::IntOp::Sub, a: one, b: eq });
-                return Some(dst);
-            }
-            self.ops.truncate(ops_mark);
-            self.live_heap_handles.truncate(lhh_mark);
-        }
-        None
-    }
-
-    /// Extracted from `Self::lower_scalar_binop` (seventh-round split, cog reduction):
-    /// the short-circuit `and`/`or` control-flow lowering + the final eager `IntBinOp`
-    /// fallback (with the narrow signed-division-overflow guard), verbatim.
-    fn lower_scalar_binop_shortcircuit_or_int(
-        &mut self,
-        op: &almide_ir::BinOp,
-        left: &IrExpr,
-        right: &IrExpr,
-    ) -> Option<ValueId> {
-        use almide_ir::BinOp;
-        // The short-circuit `and`/`or` sub-chain ALWAYS returns (`Some` on success, `None`
-        // as an explicit wall) whenever its own guard is true — it never falls through to
-        // the int-op chain below (see the original `return None;` at its guard's tail).
-        // Re-checking the same pure guard here (no side effects, so evaluating it twice is
-        // safe) lets the router pick the right helper without new shared state.
-        if matches!(op, BinOp::And | BinOp::Or) && matches!(left.ty, Ty::Bool) {
-            return self.lower_scalar_binop_shortcircuit(op, left, right);
-        }
-        self.lower_scalar_binop_int_fallback(op, left, right)
-    }
-
-    /// Extracted from `Self::lower_scalar_binop_shortcircuit_or_int` (eighth-round split,
-    /// cog reduction): the short-circuit `and`/`or` control-flow lowering, verbatim (only
-    /// called when the caller has already confirmed the `and`/`or`-over-Bool guard).
-    fn lower_scalar_binop_shortcircuit(
-        &mut self,
-        op: &almide_ir::BinOp,
-        left: &IrExpr,
-        right: &IrExpr,
-    ) -> Option<ValueId> {
-        use almide_ir::BinOp;
-        // SHORT-CIRCUIT `and`/`or` — native AND the interp oracle evaluate the RHS LAZILY
-        // (only when the LHS does not already decide the result). The prior EAGER `IntOp::And`/
-        // `Or` (materializing BOTH operands) made a RHS with a trap/side effect (`a != 0 and
-        // (10 / a) > 0`, `len > 5 and xs[5] == 0`) execute unconditionally → a divide-by-zero /
-        // OOB-`elem_addr` trap native never reaches. Lower to control flow so the RHS ops are
-        // emitted INSIDE the taken branch only:
-        //   `a and b` → `if a then b else false`   (RHS only when a is true)
-        //   `a or  b` → `if a then true else b`    (RHS only when a is false)
-        // Uses the same IfThen/Else/EndIf scalar markers as `try_lower_scalar_if`; the LHS is a
-        // pure Bool scalar, so no per-arm heap frame is needed. A non-lowerable operand rolls
-        // back (truncate) and falls through to the deferred path — never both-arms, never wrong.
-        if matches!(op, BinOp::And | BinOp::Or) && matches!(left.ty, Ty::Bool) {
-            let ops_mark = self.ops.len();
-            let lhh_mark = self.live_heap_handles.len();
-            // The RHS is evaluated INSIDE the taken IfThen/Else branch, so use
-            // `lower_scalar_operand` — it wraps the operand in a per-branch frame that frees any
-            // transient heap temp it allocates (a `contains(y, "@")` materializes its String
-            // arg) WITHIN the branch, keeping it `i…d`-balanced. (The eager path used
-            // `lower_scalar_operand` too; using bare `lower_scalar_value` walled those heap-temp
-            // operands → a coverage regression.) The LHS (a pure Bool) is likewise framed.
-            if let Some(lhs) = self.lower_scalar_operand(left) {
-                let dst = self.fresh_value();
-                self.ops.push(Op::IfThen { cond: lhs, dst: Some(dst) });
-                // THEN branch: `and` evaluates RHS here; `or` yields the constant `true`.
-                let then_val = if matches!(op, BinOp::And) {
-                    self.lower_scalar_operand(right)
-                } else {
-                    let t = self.fresh_value();
-                    self.ops.push(Op::ConstInt { dst: t, value: 1 });
-                    Some(t)
-                };
-                if let Some(tv) = then_val {
-                    self.ops.push(Op::Else { val: Some(tv) });
-                    // ELSE branch: `and` yields the constant `false`; `or` evaluates RHS here.
-                    let else_val = if matches!(op, BinOp::And) {
-                        let f = self.fresh_value();
-                        self.ops.push(Op::ConstInt { dst: f, value: 0 });
-                        Some(f)
-                    } else {
-                        self.lower_scalar_operand(right)
-                    };
-                    if let Some(ev) = else_val {
-                        self.ops.push(Op::EndIf { val: Some(ev) });
-                        return Some(dst);
-                    }
-                }
-            }
-            self.ops.truncate(ops_mark);
-            self.live_heap_handles.truncate(lhh_mark);
-            return None;
-        }
-        None
-    }
-
-    /// Extracted from `Self::lower_scalar_binop_shortcircuit_or_int` (eighth-round split,
-    /// cog reduction): the final eager `IntBinOp` fallback (with the narrow
-    /// signed-division-overflow guard), verbatim (only reached when the operator is NOT
-    /// `and`/`or` over Bool).
-    /// Extracted from `Self::lower_scalar_binop_int_fallback` (ninth-round split, cog
-    /// reduction): the pure `BinOp` + operand-shape → `IntOp` lookup, verbatim (a static
-    /// value computation, no `&mut self` needed).
-    fn scalar_binop_int_op(op: &almide_ir::BinOp, left_ty: &Ty) -> Option<crate::IntOp> {
-        use almide_ir::BinOp;
-        Some(match op {
-            BinOp::AddInt => crate::IntOp::Add,
-            BinOp::SubInt => crate::IntOp::Sub,
-            BinOp::MulInt => crate::IntOp::Mul,
-            BinOp::DivInt => crate::IntOp::Div,
-            BinOp::ModInt => crate::IntOp::Mod,
-            // Ordering comparisons (the `if` condition) — INT or BOOL operands (Bool is an i64
-            // 0/1, and v0's bool Ord is false < true = 0 < 1, so the i64 compare is bit-exact).
-            // A Float compare uses the prim float floor above; String ordering is the cmp-call
-            // above. Gate on the operand type.
-            BinOp::Lt if Self::int_ord_operand_ty(left_ty) => crate::IntOp::Lt,
-            BinOp::Lte if Self::int_ord_operand_ty(left_ty) => crate::IntOp::Le,
-            BinOp::Gt if Self::int_ord_operand_ty(left_ty) => crate::IntOp::Gt,
-            BinOp::Gte if Self::int_ord_operand_ty(left_ty) => crate::IntOp::Ge,
-            // Equality — INT or BOOL operands. A `Bool` is an i64 0/1 (a Var loads
-            // its 0/1, a `LitBool` materializes `ConstInt 0/1` above), so the SAME
-            // `IntOp::Eq`/`Ne` render is bit-exact for `b == false` / `b1 != b2` as
-            // for `n == 0`. (Ordering on Bool is undefined in v0, so it is NOT
-            // admitted; a Float/String/compound `==` still needs a distinct op.)
-            BinOp::Eq if Self::int_eq_operand_ty(left_ty) => crate::IntOp::Eq,
-            BinOp::Neq if Self::int_eq_operand_ty(left_ty) => crate::IntOp::Ne,
-            // (Logical `and`/`or` are SHORT-CIRCUITED via control flow above — they never
-            // reach this eager `IntBinOp` path. Native + interp evaluate the RHS lazily.)
-            // Pow, Float, concat, non-Int/Bool compares: defer.
-            _ => return None,
-        })
-    }
-
-    fn lower_scalar_binop_int_fallback(
-        &mut self,
-        op: &almide_ir::BinOp,
-        left: &IrExpr,
-        right: &IrExpr,
-    ) -> Option<ValueId> {
-        let iop = Self::scalar_binop_int_op(op, &left.ty)?;
-        let a = self.lower_scalar_value(left)?;
-        let b = self.lower_scalar_value(right)?;
-        self.emit_narrow_div_overflow_guard(iop, &left.ty, a, b);
-        let dst = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst, op: iop, a, b });
-        Some(dst)
-    }
-
-    /// Extracted from `Self::lower_scalar_binop_int_fallback` (ninth-round split, cog
-    /// reduction): the narrow signed-division-overflow guard injection, verbatim.
-    fn emit_narrow_div_overflow_guard(&mut self, iop: crate::IntOp, left_ty: &Ty, a: ValueId, b: ValueId) {
-        // NARROW signed division overflow (`Int8` MIN ÷ -1 — int8_div_overflow):
-        // the operands live in the i64 model, so the preamble's checked helper
-        // only catches i64::MIN ÷ -1; the narrow MIN wraps silently (v0 aborts
-        // "Error: integer overflow" + exit 1). Inject the width guard as MIR ops:
-        // if (a == MIN_w) & (b == -1) → prim.die with the SAME message bytes.
-        if !matches!(iop, crate::IntOp::Div | crate::IntOp::Mod) {
-            return;
-        }
-        let min_w = match left_ty {
-            Ty::Int8 => Some(-128i64),
-            Ty::Int16 => Some(-32768i64),
-            Ty::Int32 => Some(-2147483648i64),
-            _ => None,
-        };
-        let Some(mw) = min_w else { return };
-        let minc = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: minc, value: mw });
-        let negc = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: negc, value: -1 });
-        let c1 = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: c1, op: crate::IntOp::Eq, a, b: minc });
-        let c2 = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: c2, op: crate::IntOp::Eq, a: b, b: negc });
-        let both = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: both, op: crate::IntOp::And, a: c1, b: c2 });
-        let msg = self.fresh_value();
-        self.ops.push(Op::Alloc {
-            dst: msg,
-            repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-            init: crate::Init::Str("Error: integer overflow\n".into()),
-        });
-        let mh = self.fresh_value();
-        self.ops.push(Op::Prim { kind: crate::PrimKind::Handle, dst: Some(mh), args: vec![msg] });
-        self.ops.push(Op::IfThen { cond: both, dst: None });
-        self.ops.push(Op::Prim { kind: crate::PrimKind::Die, dst: None, args: vec![mh] });
-        self.ops.push(Op::Else { val: None });
-        self.ops.push(Op::EndIf { val: None });
-        // the message block is dead on the non-abort path — release it
-        self.ops.push(Op::Drop { v: msg });
-    }
-
-    /// Lower a `prim.*` PRIMITIVE-FLOOR call to an [`Op::Prim`] — the v1 self-host
-    /// floor (raw memory + the fd_write host call), mapped by name, NOT a real
-    /// `CallFn`/runtime symbol. Each arg lowers to a ValueId via
-    /// [`Self::lower_scalar_value`] (a handle var / int literal / int-arith). Returns
-    /// the result `dst` (load / fd_write / handle) or `None` (a store is Unit).
-    pub(crate) fn lower_prim_call(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        // Each `matches!` mirrors the exact `func == "…"` name set the corresponding
-        // group's own blocks test — a router by NAME only, no behavior of its own; the
-        // group function is never called for a name outside its set.
-        if matches!(
-            func,
-            "alloc_str" | "alloc_bytes" | "alloc_list" | "alloc_list_f64" | "alloc_set"
-                | "alloc_map" | "alloc_value" | "alloc_list_str" | "alloc_set_str"
-                | "alloc_map_str" | "alloc_map_skv" | "alloc_map_kv" | "store_str"
-        ) {
-            return self.lower_prim_call_alloc(func, args);
-        }
-        if matches!(
-            func,
-            "args_get_list" | "args_get_list_full" | "env_get" | "read_text_file"
-                | "read_bytes_file" | "read_dir" | "write_text_file" | "make_dir"
-                | "remove_all" | "path_filestat" | "path_exists"
-        ) {
-            return self.lower_prim_call_fs_env(func, args);
-        }
-        if matches!(func, "ptr_to_int" | "int_to_ptr" | "read_line" | "read_n_bytes") {
-            return self.lower_prim_call_ptr_io(func, args);
-        }
-        self.lower_prim_call_generic(func, args)
-    }
-
-    /// Extracted from `Self::lower_prim_call` (eleventh-round split, cog reduction): the
-    /// `alloc_*`/`store_str` name group, verbatim (only ever called for a name in the
-    /// router's matching `matches!` set, so no "unrecognized name" fallthrough is needed).
-    fn lower_prim_call_alloc(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        if matches!(func, "alloc_str" | "alloc_bytes" | "alloc_list" | "alloc_list_f64" | "alloc_set" | "alloc_map" | "alloc_value") {
-            return self.lower_prim_call_alloc_scalar(func, args);
-        }
-        self.lower_prim_call_alloc_str(func, args)
-    }
-
-    /// Extracted from `Self::lower_prim_call_alloc` (twelfth-round split, cog reduction):
-    /// the scalar-element alloc name group, verbatim (only ever called for a name in the
-    /// caller's matching `matches!` set).
-    fn lower_prim_call_alloc_scalar(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        // `prim.alloc_str(byte_len)` allocates a runtime-sized OWNED String — an `Op::Alloc`
-        // (cert `i`, a fresh owned object), NOT a scalar prim. The caller fills its bytes
-        // via `prim.store8`; the result is moved out / dropped like any heap value.
-        // `prim.alloc_str(n)` / `prim.alloc_bytes(n)` BOTH allocate a runtime-sized OWNED byte
-        // block (`Init::DynStr`: rc=1, len set, data filled by store8) — physically identical;
-        // they differ only in the prim's DECLARED return type (String vs Bytes). A flat heap
-        // value (no nested ownership), moved out / dropped like any String.
-        if func == "alloc_str" || func == "alloc_bytes" {
-            let len_v = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported(format!("prim.{func} length is not a lowerable scalar"))
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Alloc {
-                dst,
-                repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-                init: crate::Init::DynStr { len: len_v },
-            });
-            return Ok(Some(dst));
-        }
-        // `prim.alloc_list(n)` allocates a runtime-sized OWNED `List[Int]` of n i64 slots —
-        // an `Op::Alloc` (cert `i`), the list-building sibling of alloc_str. The caller
-        // fills it via `prim.store64`; moved out / dropped like any heap value.
-        if func == "alloc_list" || func == "alloc_list_f64" || func == "alloc_set" || func == "alloc_map" || func == "alloc_value" {
-            let len_v = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.alloc_list length is not a lowerable scalar".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Alloc {
-                dst,
-                repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-                init: crate::Init::DynList { len: len_v },
-            });
-            return Ok(Some(dst));
-        }
-        unreachable!("lower_prim_call_alloc_scalar called with a name outside its caller-matched set: {func}")
-    }
-
-    /// Extracted from `Self::lower_prim_call_alloc` (twelfth-round split, cog reduction):
-    /// the heap-element (nested-ownership) alloc + `store_str` name group, verbatim (only
-    /// ever called for a name outside `lower_prim_call_alloc_scalar`'s set).
-    fn lower_prim_call_alloc_str(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        use crate::PrimKind;
-        // `prim.alloc_list_str(n)` allocates a runtime-sized OWNED `List[String]` (n slots,
-        // physically identical to alloc_list) — but the dst is tracked as a NESTED-OWNERSHIP
-        // list, so its scope-end drop is a recursive `DropListStr` (frees the owned element
-        // Strings) and `prim.store_str` Consumes each String moved into it (Machinery 2).
-        if func == "alloc_list_str" || func == "alloc_set_str" || func == "alloc_map_str" || func == "alloc_map_skv" || func == "alloc_map_kv" {
-            let len_v = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.alloc_list_str length is not a lowerable scalar".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Alloc {
-                dst,
-                repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-                init: crate::Init::DynListStr { len: len_v },
-            });
-            self.heap_elem_lists.insert(dst);
-            return Ok(Some(dst));
-        }
-        // `prim.store_str(list, byte_addr_of_slot, piece)` — store the String `piece`'s handle
-        // into the list slot at `byte_addr_of_slot` AND CONSUME the piece (its reference is
-        // MOVED into the list, which now owns it — cert `m`, removed from the scope drop set).
-        // The slot holds the i64-widened handle; `DropListStr` later i32.wrap's it to free.
-        if func == "store_str" {
-            let addr = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.store_str slot address is not a lowerable scalar".into())
-            })?;
-            // The piece must be a tracked heap var (so we can Consume it). Its handle value:
-            let piece = match &args[1].kind {
-                IrExprKind::Var { id } => self.value_for(*id)?,
-                _ => {
-                    return Err(LowerError::Unsupported(
-                        "prim.store_str piece must be a heap variable (to consume)".into(),
-                    ))
-                }
-            };
-            // The slot value is the piece's HANDLE (its address as an i64). Op::Prim Handle
-            // gives that; store it 8-wide at the slot, then Consume the piece (move-out).
-            let handle = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(handle), args: vec![piece] });
-            self.ops.push(Op::Prim { kind: PrimKind::Store { width: 8 }, dst: None, args: vec![addr, handle] });
-            self.ops.push(Op::Consume { v: piece });
-            self.live_heap_handles.retain(|h| *h != piece);
-            return Ok(None);
-        }
-        unreachable!("lower_prim_call_alloc called with a name outside its router-matched set: {func}")
-    }
-
-    /// Extracted from `Self::lower_prim_call` (eleventh-round split, cog reduction): the
-    /// WASI env/fs-floor name group, verbatim (only ever called for a name in the
-    /// router's matching `matches!` set, so no "unrecognized name" fallthrough is needed).
-    fn lower_prim_call_fs_env(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        if matches!(func, "args_get_list" | "args_get_list_full" | "env_get") {
-            return self.lower_prim_call_env(func, args);
-        }
-        self.lower_prim_call_fs(func, args)
-    }
-
-    /// Extracted from `Self::lower_prim_call_fs_env` (twelfth-round split, cog
-    /// reduction): the CLI-args/environ name group, verbatim (only ever called for a
-    /// name in the caller's matching `matches!` set).
-    fn lower_prim_call_env(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        use crate::PrimKind;
-        // `prim.args_get_list()` — the WASI args→`List[String]` floor (env.args). NO
-        // args; its dst is a FRESH OWNED `List[String]` of `argv[1..]` (a heap Ptr, like
-        // an Alloc), so it is registered like `alloc_list_str`: a NESTED-OWNERSHIP list
-        // whose scope-end drop is the recursive `DropListStr` (frees the owned element
-        // Strings) — a flat `Drop` would leak them. Carries Capability::CliArgs (counted
-        // in cap_witness). The render emits the WASI args_sizes_get/args_get sequence.
-        if func == "args_get_list" {
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::ArgsGetList, dst: Some(dst), args: vec![] });
-            self.heap_elem_lists.insert(dst);
-            return Ok(Some(dst));
-        }
-        // `prim.args_get_list_full()` — the argv[0]-INCLUSIVE twin (process.args =
-        // std::env::args()). Same fresh OWNED List[String] + DropListStr + CliArgs
-        // discipline; renders through the SAME parameterized $args_get_list bridge.
-        if func == "args_get_list_full" {
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::ArgsGetListFull, dst: Some(dst), args: vec![] });
-            self.heap_elem_lists.insert(dst);
-            return Ok(Some(dst));
-        }
-        // `prim.env_get(name)` — the WASI environ lookup floor (env.get). ONE BORROWED
-        // `String` arg (the variable name; the caller still owns it). Its dst is a FRESH
-        // OWNED `Option[String]` in the `materialize_opt_str_some` layout (0-slot none /
-        // 1-slot some owning the value String @12), registered in `heap_elem_lists` so
-        // the scope-end drop is the flat `DropListStr` (frees the payload String, if
-        // any, then the block). Carries Capability::CliArgs — the Env profile's
-        // canonical cap (counted in cap_witness exactly like ArgsGetList).
-        if func == "env_get" && args.len() == 1 {
-            let key = match self.lower_call_args(args)?.into_iter().next() {
-                Some(CallArg::Handle(v)) => v,
-                _ => {
-                    return Err(LowerError::Unsupported(
-                        "prim.env_get needs a borrowed String name".into(),
-                    ))
-                }
-            };
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::EnvGet, dst: Some(dst), args: vec![key] });
-            self.heap_elem_lists.insert(dst);
-            return Ok(Some(dst));
-        }
-        // `env_get` with a wrong arg count (the ONLY name in this group with an extra
-        // guard beyond the bare name test) falls all the way through the original
-        // single-match to the terminal "unknown primitive" wall — replicated verbatim
-        // here (NOT `unreachable!`: this guard genuinely can fail for a matched name).
-        Err(LowerError::Unsupported(format!("unknown primitive prim.{func}")))
-    }
-
-    /// Extracted from `Self::lower_prim_call_fs_env` (twelfth-round split, cog
-    /// reduction): the WASI filesystem-floor name group, verbatim (only ever called
-    /// for a name in the caller's matching `matches!` set).
-    fn lower_prim_call_fs(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        use crate::PrimKind;
-        // `prim.read_text_file(path)` — the WASI file-read floor (fs.read_text). ONE
-        // BORROWED `String` arg (the path; the caller still owns it). Its dst is a FRESH
-        // OWNED `Result[String, String]` built by the render in the EXACT
-        // `materialize_result_str` cap-as-tag layout (1-slot DynListStr, payload @12, tag
-        // @16). Tracked like a heap-Ok Result: `materialized_results_str` so a downstream
-        // `match`/`!` reads tag @16, AND `heap_elem_lists` so the heap-payload bind gates open
-        // AND the scope-end drop is the flat `DropListStr` (frees the one owned String @12 +
-        // the block — a flat `Drop` would leak the String). Carries Capability::FsRead
-        // (counted in cap_witness). The render emits the WASI path_open/fd_read sequence.
-        if func == "read_text_file" || func == "read_bytes_file" {
-            // read_bytes_file is the raw-bytes twin: the SAME WASI floor + Result block
-            // (the render's $read_text_file reads raw bytes; only the almd-level Ok TYPE
-            // differs), so one PrimKind serves both.
-            let path = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.read_text_file path is not a lowerable scalar/handle".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::ReadTextFile, dst: Some(dst), args: vec![path] });
-            self.materialized_results_str.insert(dst);
-            self.heap_elem_lists.insert(dst);
-            return Ok(Some(dst));
-        }
-        // `prim.read_dir(path)` — the WASI directory-listing floor (fs.list_dir). ONE BORROWED
-        // `String` arg (the path). Its dst is a FRESH OWNED `Result[List[String], String]` built
-        // by the render ($read_dir) in the cap-as-tag layout (1-slot wrapper, payload @12 = a
-        // List[String], tag @16). Tracked like a heap-Ok Result: `materialized_results_str` so a
-        // downstream `match`/`!` reads tag @16, AND `heap_elem_lists` so the heap-payload bind
-        // gates open, AND `list_str_result_results` so the scope-end drop is the RECURSIVE
-        // `DropResultListStr` (frees the payload List's element Strings + block; a flat
-        // `DropListStr` would leak them) — checked BEFORE heap_elem_lists in `drop_op_for`.
-        // Carries Capability::FsRead (counted in cap_witness). The render emits the WASI
-        // path_open(O_DIRECTORY)/fd_readdir sequence (skip `.`/`..`, sort, build the list).
-        if func == "read_dir" {
-            let path = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.read_dir path is not a lowerable scalar/handle".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim { kind: PrimKind::ReadDir, dst: Some(dst), args: vec![path] });
-            self.materialized_results_str.insert(dst);
-            self.heap_elem_lists.insert(dst);
-            self.list_str_result_results.insert(dst);
-            return Ok(Some(dst));
-        }
-        // `prim.write_text_file(path, content)` — the WASI file-WRITE floor (fs.write). TWO
-        // BORROWED `String` args (the path + the content; the caller still owns both). Its dst is a
-        // FRESH OWNED `Result[Unit, String]` built by the render ($write_text_file): Ok(()) with
-        // `len@4 = 0` (no payload String — the `materialize_result_ok` convention) so the scope-end
-        // flat `DropListStr` frees nothing at @12, or Err(msg) with `len@4 = 1` + `@12 = msg` (the
-        // flat drop frees the one owned message). Tracked like a heap Result: `materialized_results_str`
-        // so a downstream `match`/`!` reads the @16 tag, AND `heap_elem_lists` so the heap-payload
-        // bind gates open AND the scope-end drop is the flat `DropListStr` (sound for BOTH arms given
-        // the `len@4 = 0` Ok convention — NO `list_str_result_results`: there is no nested payload).
-        // Carries Capability::FsWrite (counted in cap_witness). The render emits the WASI
-        // path_open(O_CREAT|O_TRUNC)/fd_write sequence.
-        if func == "write_text_file" {
-            let path = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.write_text_file path is not a lowerable scalar/handle".into())
-            })?;
-            let content = self.lower_scalar_value(&args[1]).ok_or_else(|| {
-                LowerError::Unsupported("prim.write_text_file content is not a lowerable scalar/handle".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim {
-                kind: PrimKind::WriteTextFile,
-                dst: Some(dst),
-                args: vec![path, content],
-            });
-            self.materialized_results_str.insert(dst);
-            self.heap_elem_lists.insert(dst);
-            return Ok(Some(dst));
-        }
-        // `prim.make_dir(path)` — the WASI directory-CREATE floor (fs.mkdir_p). ONE BORROWED
-        // `String` arg (the path; the caller still owns it). Its dst is a FRESH OWNED
-        // `Result[Unit, String]` built by the render ($make_dir): Ok(()) with `len@4 = 0` (no
-        // payload String — the `materialize_result_ok` convention, IDENTICAL to write_text_file's
-        // Ok arm) so the scope-end flat `DropListStr` frees nothing at @12, or Err(msg) with
-        // `len@4 = 1` + `@12 = msg` (the flat drop frees the one owned message). Tracked exactly
-        // like write_text_file's heap Result: `materialized_results_str` so a downstream `match`/`!`
-        // reads the @16 tag, AND `heap_elem_lists` so the heap-payload bind gates open AND the
-        // scope-end drop is the flat `DropListStr`. Carries Capability::FsWrite (a mkdir IS a
-        // filesystem write — counted in cap_witness). The render emits the WASI recursive
-        // path_create_directory sequence.
-        if func == "make_dir" {
-            let path = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.make_dir path is not a lowerable scalar/handle".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim {
-                kind: PrimKind::MakeDir,
-                dst: Some(dst),
-                args: vec![path],
-            });
-            self.materialized_results_str.insert(dst);
-            self.heap_elem_lists.insert(dst);
-            return Ok(Some(dst));
-        }
-        // `prim.remove_all(path)` — the WASI recursive-remove floor (fs.remove_all). ONE BORROWED
-        // `String` arg (the path; the caller still owns it). Its dst is a FRESH OWNED
-        // `Result[Unit, String]` built by the render ($remove_all): Ok(()) with `len@4 = 0` (no
-        // payload String — the `materialize_result_ok` convention, IDENTICAL to make_dir's Ok arm)
-        // so the scope-end flat `DropListStr` frees nothing at @12, or Err(msg) with `len@4 = 1` +
-        // `@12 = msg` (the flat drop frees the one owned message). Tracked exactly like make_dir's
-        // heap Result: `materialized_results_str` so a downstream `match`/`!` reads the @16 tag, AND
-        // `heap_elem_lists` so the heap-payload bind gates open AND the scope-end drop is the flat
-        // `DropListStr`. Carries Capability::FsWrite (a recursive remove IS a filesystem write —
-        // counted in cap_witness). The render emits the WASI recursive
-        // path_remove_directory/path_unlink_file sequence.
-        if func == "remove_all" {
-            let path = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.remove_all path is not a lowerable scalar/handle".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim {
-                kind: PrimKind::RemoveAll,
-                dst: Some(dst),
-                args: vec![path],
-            });
-            self.materialized_results_str.insert(dst);
-            self.heap_elem_lists.insert(dst);
-            return Ok(Some(dst));
-        }
-        // `prim.path_exists(path)` — the WASI path-stat floor (fs.exists). ONE BORROWED `String`
-        // arg (the path; the caller still owns it). Its dst is a SCALAR `Bool` (i64 0/1) — UNLIKE
-        // every other fs prim, a stat allocates NO heap result, so the dst is tracked in NO
-        // classification set (no `materialized_results_str` / `heap_elem_lists`): it is a plain
-        // scalar with no scope-end drop and no ownership-cert `i`. Carries Capability::FsRead (a
-        // stat IS a filesystem read — counted in cap_witness). The render emits the WASI
-        // path_filestat_get query (errno 0 = exists).
-        // `prim.path_filestat(bufaddr, path)` — the WASI FULL-stat floor (fs.stat). TWO args: a
-        // raw scratch ADDRESS (an i64 scalar — the self-host's own Bytes data region, so the
-        // caller owns the buffer) and a BORROWED `String` path. dst = the SCALAR errno (0 = the
-        // 64-byte WASI filestat is at bufaddr). Like path_exists this allocates NO heap result —
-        // the dst joins no classification set. Carries Capability::FsRead (counted in cap_witness).
-        if func == "path_filestat" {
-            let bufaddr = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported(
-                    "prim.path_filestat buffer address is not a lowerable scalar".into(),
-                )
-            })?;
-            let path = self.lower_scalar_value(&args[1]).ok_or_else(|| {
-                LowerError::Unsupported(
-                    "prim.path_filestat path is not a lowerable scalar/handle".into(),
-                )
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim {
-                kind: PrimKind::PathFilestat,
-                dst: Some(dst),
-                args: vec![bufaddr, path],
-            });
-            return Ok(Some(dst));
-        }
-        if func == "path_exists" {
-            let path = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported("prim.path_exists path is not a lowerable scalar/handle".into())
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim {
-                kind: PrimKind::PathExists,
-                dst: Some(dst),
-                args: vec![path],
-            });
-            return Ok(Some(dst));
-        }
-        unreachable!("lower_prim_call_fs called with a name outside its caller-matched set: {func}")
-    }
-
-    /// Extracted from `Self::lower_prim_call` (eleventh-round split, cog reduction): the
-    /// pointer-cast/stdin-read name group, verbatim (only ever called for a name in the
-    /// router's matching `matches!` set, so no "unrecognized name" fallthrough is needed).
-    fn lower_prim_call_ptr_io(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        use crate::PrimKind;
-        // `prim.ptr_to_int` / `prim.int_to_ptr` — REINTERPRET casts (identity at the
-        // value level: the RawPtr IS the i64 address). No op emitted — the operand's
-        // ValueId passes through, so the cert sees nothing (a pure hat-swap).
-        if func == "ptr_to_int" || func == "int_to_ptr" {
-            let v = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported(
-                    "prim ptr cast operand is not a lowerable scalar".into(),
-                )
-            })?;
-            return Ok(Some(v));
-        }
-        // `prim.read_line()` — the WASI stdin-line floor (io.read_line). NO args. Its dst is a
-        // FRESH OWNED canonical `String` (one line of stdin, newline excluded) built by the render
-        // ($read_line). A plain String owns NO nested handles, so it is tracked in NO classification
-        // set — its scope-end drop (if not moved out as a return) is the flat `Op::Drop` that frees
-        // the block (a `DropListStr` would WRONGLY treat the byte payload as i64 element handles).
-        // Carries Capability::Stdin (counted in cap_witness). The render emits the byte-by-byte
-        // fd_read-from-fd-0 sequence.
-        if func == "read_line" {
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim {
-                kind: PrimKind::ReadLine,
-                dst: Some(dst),
-                args: vec![],
-            });
-            return Ok(Some(dst));
-        }
-        // `prim.read_n_bytes(n)` — the WASI stdin-N-bytes floor (io.read_n_bytes). The n arg is a
-        // scalar Int (byte count); dst is a FRESH OWNED `Bytes` block (byte-buffer layout, built by the
-        // preamble `$read_n_bytes`). Carries Capability::Stdin (counted via certificate.rs). Like
-        // read_line, a plain Bytes owns no nested handles, so its scope-end drop is the flat `Op::Drop`.
-        if func == "read_n_bytes" {
-            let n = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-                LowerError::Unsupported(
-                    "prim.read_n_bytes needs a scalar Int byte count not in this brick".into(),
-                )
-            })?;
-            let dst = self.fresh_value();
-            self.ops.push(Op::Prim {
-                kind: PrimKind::ReadNBytes,
-                dst: Some(dst),
-                args: vec![n],
-            });
-            return Ok(Some(dst));
-        }
-        unreachable!("lower_prim_call_ptr_io called with a name outside its router-matched set: {func}")
-    }
-
-    /// Extracted from `Self::lower_prim_call_generic` (twelfth-round split, cog
-    /// reduction): the load/store/rc/misc-floor half of the name → `PrimKind` table,
-    /// verbatim (only ever called for a name in the caller's matching `matches!` set).
-    fn prim_kind_structural(&self, func: &str) -> Result<crate::PrimKind, LowerError> {
-        if matches!(func, "handle" | "die" | "load8" | "load32" | "load64" | "load_str" | "load_handle" | "store32" | "store8" | "store64") {
-            return Ok(Self::prim_kind_load_store(func));
-        }
-        self.prim_kind_rc_io(func)
-    }
-
-    /// Extracted from `Self::prim_kind_structural` (thirteenth-round split, cog
-    /// reduction): the handle/die/load/store sub-table, verbatim (a pure lookup, only
-    /// ever called for a name in the caller's matching `matches!` set).
-    fn prim_kind_load_store(func: &str) -> crate::PrimKind {
-        use crate::PrimKind;
-        match func {
-            "handle" => PrimKind::Handle,
-            "die" => PrimKind::Die,
-            "load8" => PrimKind::Load { width: 1 },
-            "load32" => PrimKind::Load { width: 4 },
-            "load64" => PrimKind::Load { width: 8 },
-            // Load a 4-byte handle KEEPING Ptr repr — reads a String element out of a list slot
-            // (a borrow of the slot's String, for passing to a closure / String fn).
-            "load_str" => PrimKind::LoadHandle,
-            // Generic typed `load_handle[A]` — the same i32-handle-keeping load as `load_str`, for
-            // reading a `List[Value]`/`Value` payload out of a Value's slot (the Value model floor).
-            "load_handle" => PrimKind::LoadHandle,
-            "store32" => PrimKind::Store { width: 4 },
-            "store8" => PrimKind::Store { width: 1 },
-            "store64" => PrimKind::Store { width: 8 },
-            _ => unreachable!("prim_kind_load_store called with a name outside its caller-matched set: {func}"),
-        }
-    }
-
-    /// Extracted from `Self::prim_kind_structural` (thirteenth-round split, cog
-    /// reduction): the rc_dec/rc_inc/fd_write/random/clock sub-table, verbatim (only
-    /// ever called for a name outside `prim_kind_load_store`'s set).
-    fn prim_kind_rc_io(&self, func: &str) -> Result<crate::PrimKind, LowerError> {
-        use crate::PrimKind;
-        Ok(match func {
-            // Raw refcount free/acquire — the Value drop/copy mechanism. GATED to the value-model
-            // self-host fns (the trusted recursive-free / shallow-copy, like the inline DropListStr):
-            // an UNTRACKED free exposed to arbitrary code would let any fn double-free outside the
-            // ownership cert's sight, so only the value-model drop/copy routines may name it: the
-            // recursive drop (`__drop_value`, rc_dec), the array shallow-copy (`__varr_copy`, rc_inc),
-            // the as_array element-list fill (`__vfill`, rc_inc), and the heap-element list-concat copy
-            // (`__lc_copy_rc`, rc_inc — the new list co-owns each appended element, balanced by the
-            // source's recursive DropListStr/DropListValue). See docs/roadmap/active/v1-value-model.md.
-            //
-            // TRUST GROUNDING (柱C Brick 3): these names are a CO-OWN-PRODUCER / RECURSIVE-DROP whitelist
-            // — a producer (`__varr_copy`/`__vobj_fill`/`__copy_value`/`__lc_copy_rc`/…) rc_inc's each
-            // loaded element (+1) into a fresh container; its balancing rc_dec lives in the SEPARATE
-            // recursive drop (`__drop_value`/`__vdrop_arr`/…) over the SAME elements. That cross-loop,
-            // element-count-keyed balance is PROVEN leak/double-free-free on the Coq kernel by
-            // proofs/CoownLoop.v (`coown_fill_drop_neutral` ⇒ `coown_copy_no_leak` + `…no_double_free`,
-            // in the check.sh proof gate). So this gate is no longer bare trust: a name belongs here iff
-            // it is a co-own producer or recursive-drop consumer following that proven pattern, and its
-            // adherence is ratcheted by the spec/wasm_cross/*_leak_loop fixtures. Cert-PROVING each
-            // producer per-function (retiring the whitelist) needs the typed nested-element model + the
-            // cross-function fill↔drop pairing that consumes CoownLoop.v — the remaining Brick-3
-            // engineering (docs/roadmap/active/value-rc-cert.md).
-            // The co-own producer / recursive-drop whitelist lives in ONE shared anchor
-            // (crate::coown_names) grounded in proofs/CoownLoop.v + CoownCompose.v — see that module.
-            "rc_dec" | "rc_inc"
-                if crate::coown_names::is_coown_rc_routine(self.fn_name.as_str())
-                    || self.fn_name.starts_with("__drop_")
-                    // `__krec_uniqfill_<R>` — the GENERATED list.unique fill over a
-                    // String-field-record element (C-015): rc_inc each KEPT element
-                    // into the result (the __uh_acquire pattern, per-type generated
-                    // like __drop_*; drop_sources.rs is the single emitter).
-                    || self.fn_name.starts_with("__krec_uniqfill_") =>
-            {
-                // `__drop_*` also covers the GENERATED per-type custom-variant recursive drops
-                // (`__drop_Expr`, ADT brick 5b) — the same trusted prim-only free routine.
-                if func == "rc_dec" { PrimKind::RcDec } else { PrimKind::RcInc }
-            }
-            "rc_dec" | "rc_inc" => {
-                return Err(LowerError::Unsupported(format!(
-                    "prim.{func} is restricted to the value-model drop/copy routines (untracked free)"
-                )))
-            }
-            "fd_write" => PrimKind::FdWrite,
-            "random_get" => PrimKind::RandomGet,
-            "clock_time_get" => PrimKind::ClockTimeGet,
-            _ => unreachable!("prim_kind_structural called with a name outside its caller-matched set: {func}"),
-        })
-    }
-
-    /// Extracted from `Self::lower_prim_call_generic` (twelfth-round split, cog
-    /// reduction): the FLOAT-floor half of the name → `PrimKind` table, verbatim (a
-    /// pure lookup, only ever called for a name in the caller's matching `matches!`
-    /// set — total, so no `Result`/`Option` needed).
-    fn prim_kind_float(func: &str) -> crate::PrimKind {
-        if matches!(
-            func,
-            "fabs" | "fsqrt" | "ffloor" | "fceil" | "fneg" | "fadd" | "fsub" | "fmul" | "fdiv"
-                | "fmin" | "fmax" | "fcopysign"
-        ) {
-            return Self::prim_kind_float_arith(func);
-        }
-        Self::prim_kind_float_cmp_conv(func)
-    }
-
-    /// Extracted from `Self::prim_kind_float` (thirteenth-round split, cog reduction):
-    /// the float unary/binary-arithmetic sub-table, verbatim (a pure lookup, only ever
-    /// called for a name in the caller's matching `matches!` set).
-    fn prim_kind_float_arith(func: &str) -> crate::PrimKind {
-        use crate::PrimKind;
-        match func {
-            // The FLOAT floor (the f64 bits live in the i64-uniform value; render reinterprets).
-            "fabs" => PrimKind::FloatUn(crate::FUnOp::Abs),
-            "fsqrt" => PrimKind::FloatUn(crate::FUnOp::Sqrt),
-            "ffloor" => PrimKind::FloatUn(crate::FUnOp::Floor),
-            "fceil" => PrimKind::FloatUn(crate::FUnOp::Ceil),
-            "fneg" => PrimKind::FloatUn(crate::FUnOp::Neg),
-            "fadd" => PrimKind::FloatBin(crate::FBinOp::Add),
-            "fsub" => PrimKind::FloatBin(crate::FBinOp::Sub),
-            "fmul" => PrimKind::FloatBin(crate::FBinOp::Mul),
-            "fdiv" => PrimKind::FloatBin(crate::FBinOp::Div),
-            "fmin" => PrimKind::FloatBin(crate::FBinOp::Min),
-            "fmax" => PrimKind::FloatBin(crate::FBinOp::Max),
-            "fcopysign" => PrimKind::FloatBin(crate::FBinOp::CopySign),
-            _ => unreachable!("prim_kind_float_arith called with a name outside its caller-matched set: {func}"),
-        }
-    }
-
-    /// Extracted from `Self::prim_kind_float` (thirteenth-round split, cog reduction):
-    /// the float comparison/conversion sub-table, verbatim (only ever called for a name
-    /// outside `prim_kind_float_arith`'s set).
-    fn prim_kind_float_cmp_conv(func: &str) -> crate::PrimKind {
-        use crate::PrimKind;
-        match func {
-            "flt" => PrimKind::FloatCmp(crate::FCmpOp::Lt),
-            "fle" => PrimKind::FloatCmp(crate::FCmpOp::Le),
-            "fgt" => PrimKind::FloatCmp(crate::FCmpOp::Gt),
-            "fge" => PrimKind::FloatCmp(crate::FCmpOp::Ge),
-            "feq" => PrimKind::FloatCmp(crate::FCmpOp::Eq),
-            "fne" => PrimKind::FloatCmp(crate::FCmpOp::Ne),
-            "f2i" => PrimKind::FloatToInt,
-            "i2f" => PrimKind::IntToFloat,
-            "fbits" | "ffrombits" => PrimKind::FloatBits,
-            // f32 narrowing/widening (f32 value = its 32-bit pattern in the low half of the i64).
-            "f2f32" => PrimKind::F32Demote,
-            // `f32_2f` (Float32→Float) and `bits_to_f32` (raw 32-bit pattern→Float) are the SAME
-            // f64.promote_f32 over a low-32 f32 pattern.
-            "f32_2f" | "bits_to_f32" => PrimKind::F32Promote,
-            "i2f32" => PrimKind::IntToF32,
-            "f32bits" => PrimKind::F32Bits,
-            _ => unreachable!("prim_kind_float_cmp_conv called with a name outside its caller-matched set: {func}"),
-        }
-    }
-
-    fn lower_prim_call_generic(
-        &mut self,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<Option<ValueId>, LowerError> {
-        // Bitwise binary ops lower to a scalar `Op::IntBinOp` (i64 and/or/xor/shl/shr_s),
-        // not an `Op::Prim` — the int.band/bor/bxor/bshl/bshr floor. No ownership.
-        if let Some(op) = Self::bitop_for_name(func) {
-            return self.emit_prim_bitop(func, op, args);
-        }
-        let kind = self.prim_kind_for_name(func)?;
-        self.emit_prim_call(func, args, kind)
-    }
-
-    /// Extracted from `Self::lower_prim_call_generic` (thirteenth-round split, cog
-    /// reduction): the bitwise-op name → `IntOp` lookup, verbatim (a pure lookup, no
-    /// `&self` needed).
-    fn bitop_for_name(func: &str) -> Option<crate::IntOp> {
-        match func {
-            "band" => Some(crate::IntOp::And),
-            "bor" => Some(crate::IntOp::Or),
-            "bxor" => Some(crate::IntOp::Xor),
-            "bshl" => Some(crate::IntOp::Shl),
-            "bshr" => Some(crate::IntOp::Shr),
-            "bshr_u" => Some(crate::IntOp::ShrU),
-            _ => None,
-        }
-    }
-
-    /// Extracted from `Self::lower_prim_call_generic` (thirteenth-round split, cog
-    /// reduction): the bitwise-op emission, verbatim.
-    fn emit_prim_bitop(&mut self, func: &str, op: crate::IntOp, args: &[IrExpr]) -> Result<Option<ValueId>, LowerError> {
-        let a = self.lower_scalar_value(&args[0]).ok_or_else(|| {
-            LowerError::Unsupported(format!("prim.{func} arg 0 is not a lowerable scalar"))
-        })?;
-        let b = self.lower_scalar_value(&args[1]).ok_or_else(|| {
-            LowerError::Unsupported(format!("prim.{func} arg 1 is not a lowerable scalar"))
-        })?;
-        let dst = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst, op, a, b });
-        Ok(Some(dst))
-    }
-
-    /// Extracted from `Self::lower_prim_call_generic` (thirteenth-round split, cog
-    /// reduction): the structural-vs-float name-group dispatch, verbatim.
-    fn prim_kind_for_name(&self, func: &str) -> Result<crate::PrimKind, LowerError> {
-        if matches!(
-            func,
-            "handle" | "die" | "load8" | "load32" | "load64" | "load_str" | "load_handle"
-                | "store32" | "store8" | "store64" | "rc_dec" | "rc_inc" | "fd_write"
-                | "random_get" | "clock_time_get"
-        ) {
-            return self.prim_kind_structural(func);
-        }
-        if matches!(
-            func,
-            "fabs" | "fsqrt" | "ffloor" | "fceil" | "fneg" | "fadd" | "fsub" | "fmul" | "fdiv"
-                | "fmin" | "fmax" | "fcopysign" | "flt" | "fle" | "fgt" | "fge" | "feq" | "fne"
-                | "f2i" | "i2f" | "fbits" | "ffrombits" | "f2f32" | "f32_2f" | "bits_to_f32"
-                | "i2f32" | "f32bits"
-        ) {
-            return Ok(Self::prim_kind_float(func));
-        }
-        Err(LowerError::Unsupported(format!("unknown primitive prim.{func}")))
-    }
-
-    /// Extracted from `Self::lower_prim_call_generic` (thirteenth-round split, cog
-    /// reduction): the generic arg-lowering loop + final `Op::Prim` emission, verbatim.
-    /// Extracted from `Self::emit_prim_call` (fourteenth-round split, cog reduction):
-    /// the per-arg lowering loop BODY (one iteration), verbatim.
-    fn lower_prim_call_one_arg(&mut self, func: &str, a: &IrExpr, kind: crate::PrimKind) -> Result<ValueId, LowerError> {
-        use crate::PrimKind;
-        // A STRING-LITERAL argument to `prim.handle` — the frontend's single-use
-        // let-inliner pushes `let tbl = "…"; prim.handle(tbl)` into
-        // `prim.handle("…")` (the generated case-mapping tables). Materialize
-        // the literal block exactly as its let-bound form would (owned Alloc,
-        // scope-end drop) and hand the prim its handle — the scalar-tail
-        // deferred-Const fallback was silently returning 0 as the address.
-        if matches!(kind, PrimKind::Handle) {
-            if let IrExprKind::LitStr { value } = &a.kind {
-                let dst = self.fresh_value();
-                self.ops.push(Op::Alloc {
-                    dst,
-                    repr: repr_of(&a.ty)?,
-                    init: crate::Init::Str(value.clone()),
-                });
-                self.live_heap_handles.push(dst);
-                return Ok(dst);
-            }
-            // A COMPUTED String argument (`prim.die(prim.handle("assertion failed: "
-            // + msg))` — the 2-arg assert's computed-message die): materialize the
-            // concat/interp chain to an owned block (scope-tracked, dropped at the
-            // arm/scope end like the literal above) and hand the prim its handle.
-            // Without this the whole assert's unit-if rolled back and the wall
-            // (misleadingly) named the CONDITION.
-            if matches!(
-                &a.kind,
-                IrExprKind::BinOp { op: almide_ir::BinOp::ConcatStr, .. }
-                    | IrExprKind::StringInterp { .. }
-            ) {
-                let obj = match &a.kind {
-                    IrExprKind::BinOp { .. } => self.try_lower_concat_str(a),
-                    IrExprKind::StringInterp { parts } => self.try_lower_string_interp(parts),
-                    _ => unreachable!(),
-                };
-                if let Some(obj) = obj {
-                    self.live_heap_handles.push(obj);
-                    return Ok(obj);
-                }
-            }
-        }
-        self.lower_scalar_value(a).ok_or_else(|| {
-            LowerError::Unsupported(format!("prim.{func} argument is not a lowerable scalar/handle"))
-        })
-    }
-
-    fn emit_prim_call(&mut self, func: &str, args: &[IrExpr], kind: crate::PrimKind) -> Result<Option<ValueId>, LowerError> {
-        use crate::PrimKind;
-        let mut lowered = Vec::with_capacity(args.len());
-        for a in args {
-            lowered.push(self.lower_prim_call_one_arg(func, a, kind)?);
-        }
-        let dst = if matches!(kind, PrimKind::Store { .. } | PrimKind::RcDec | PrimKind::RcInc | PrimKind::Die | PrimKind::ProcExit) {
-            None
-        } else {
-            Some(self.fresh_value())
-        };
-        // `prim.load_str` (LoadHandle) yields a BORROW of a list slot's String — the list still owns
-        // it. Mark the result BORROWED so a `let` binding does not add it to the scope-end drop set
-        // (that would double-free with the owning list's DropListStr).
-        if matches!(kind, PrimKind::LoadHandle) {
-            if let Some(d) = dst {
-                self.param_values.insert(d);
-            }
-        }
-        self.ops.push(Op::Prim { kind, dst, args: lowered });
-        Ok(dst)
-    }
-
-    /// Register a freshly-materialized call-result temp used as a call argument: a
-    /// HEAP temp is BORROWED into the call (`Handle`) and added to the scope-end
-    /// drop set (it is owned by THIS scope, not moved out, so it is released after
-    /// the call returns); a scalar temp is passed by value. A NESTED-OWNERSHIP temp
-    /// (a `List[String]` from `set.from_list(string.split(…))`, etc.) is ALSO recorded
-    /// in `heap_elem_lists` so its scope-end drop is the recursive `DropListStr` that
-    /// frees the owned element Strings — a flat `Drop` would free only the block and
-    /// LEAK the elements (per-iteration in a loop → OOM). Cert is unchanged: one `i`
-    /// (alloc) + one `d` (drop) for the temp; DropListStr vs Drop is the runtime
-    /// realization of that same single `d`.
-    pub(crate) fn materialized_call_arg(&mut self, dst: ValueId, repr: Repr, ty: &Ty) -> CallArg {
-        if repr.is_heap() {
-            self.live_heap_handles.push(dst);
-            // A `value.as_array(v) ?? []` arg temp (the materialized `??` operand) is a
-            // Result[List[Value],String] that OWNS its inner list — its drop must free the list AND
-            // its element Values RECURSIVELY (`DropResultListValue`); the flat `heap_elem_lists`
-            // fallback would only rc_dec the inner-list handle, LEAKING the element Values (a loop
-            // OOMs). Checked BEFORE is_heap_elem_list_ty, which also matches this Result type.
-            // (A Result[Value,String]'s Ok Value is CO-OWNED — value.get Dup's the object's slot, which
-            // keeps its ref — so the flat rc_dec drop is correct there; a recursive free would
-            // double-free the still-referenced slot. So only the list case is reclassified here.)
-            if crate::lower::is_result_listval_ty(ty) {
-                self.value_result_lists.insert(dst);
-            } else if crate::lower::is_list_list_str_ty(ty) {
-                self.list_list_str_lists.insert(dst);
-            } else if crate::lower::is_list_str_str_ty(ty) {
-                // `List[(String,String)]` (map.entries) arg temp — DropListStrStr frees each tuple's
-                // two Strings; the flat heap_elem_lists fallback would leak them.
-                self.str_str_elem_lists.insert(dst);
-            } else if crate::lower::is_lenlist_list_ty(ty) {
-                self.variant_drop_handles.insert(dst, "list_lenlist".to_string());
-            } else if crate::lower::is_map_fn_ty(ty) {
-                // `Map[String, <Fn>]` arg temp — `$__drop_map_mclo` frees each value via
-                // `__drop_closure` (a flat sweep would leak every captured env slot).
-                self.variant_drop_handles.insert(dst, "map_mclo".to_string());
-            } else if let Some(hname) = self.map_named_value_drop(ty) {
-                self.variant_drop_handles.insert(dst, hname);
-            } else if crate::lower::is_map_msv_ty(ty) {
-                // `Map[String, Map[String, String]]` arg temp (the inline nested-map literal
-                // fed straight to `map.get_or` — map_fold_heap_acc's r7): `$__drop_map_msv`
-                // sweeps each last-ref inner map; the flat fallback leaked the whole nested
-                // map per iteration (loop OOM).
-                self.variant_drop_handles.insert(dst, "map_msv".to_string());
-            } else if crate::lower::is_map_mlo_ty(ty) {
-                // `Map[String, List[Option[Int]]]` arg temp — `$__drop_map_mlo` (the
-                // bind-site route, mirrored; the flat fallback would leak the value lists).
-                self.variant_drop_handles.insert(dst, "map_mlo".to_string());
-            } else if let Some(rname) = (match ty {
-                Ty::Applied(almide_lang::types::constructor::TypeConstructorId::List, a)
-                    if a.len() == 1 =>
-                {
-                    self.record_or_anon_drop_type_name(&a[0])
-                }
-                _ => None,
-            }) {
-                // A `List[<recursive-drop record>]` arg temp — `$__drop_list_<R>` (the
-                // bind-site route, mirrored; the flat fallback leaked each element's
-                // String fields — the krec-unique residue).
-                self.variant_drop_handles.insert(dst, format!("list_{rname}"));
-            } else if matches!(ty,
-                Ty::Applied(almide_lang::types::constructor::TypeConstructorId::Map, a)
-                    if a.len() == 2 && matches!(a[0], Ty::String) && !is_heap_ty(&a[1]))
-            {
-                // `Map[String, <scalar>]` arg temp — the key-slot sweep (split layout, @4 = n),
-                // mirroring the bind-site fix; the flat fallback leaked every key copy.
-                self.heap_elem_lists.insert(dst);
-            } else if crate::lower::is_heap_elem_list_ty(ty) {
-                self.heap_elem_lists.insert(dst);
-            }
-            // A `Value` call-argument temp (`f(value.array([…]))`, `f(value.str(s))`) drops via the
-            // runtime-tag-dispatched `Op::DropValue` (recursive — an Array frees its element Values, a
-            // Str its String), NOT a flat `Op::Drop` (which would leak the nested payload). Without
-            // this a tag-5 Array / tag-4 Str passed as an argument leaks at the call-site scope end.
-            if crate::lower::is_value_ty(ty) {
-                self.value_handles.insert(dst);
-            }
-            // A RECORD/TUPLE call-argument temp (`f(mk(x))` — a fresh record passed by handle) drops at
-            // the call-site scope end. Without a mask it falls to a flat `Op::Drop` (rc_dec the record
-            // block only), LEAKING every heap field (the `f(mk(x))`-in-a-loop OOM). Seed its heap-slot
-            // `record_masks` (the masked drop frees the leaf fields) and, when a field is a
-            // Map/List[heap]/record/Value, route to the recursive `$__drop_<R>` via variant_drop_handles.
-            if let Some((_, tys)) = self.aggregate_field_tys(ty) {
-                let heap_slots: Vec<usize> =
-                    (0..tys.len()).filter(|&i| is_heap_ty(&tys[i])).collect();
-                self.record_masks.insert(dst, heap_slots);
-                if let Some(name) = self.record_drop_type_name(ty) {
-                    self.variant_drop_handles.insert(dst, name);
-                }
-            }
-            CallArg::Handle(dst)
-        } else {
-            CallArg::Scalar(dst)
-        }
-    }
 }
